refactor(api): split snapshot-builder.ts to satisfy 700-line file limit
Extracts buildPressureSensitivitySnapshotOutput and its private helpers
to snapshot-compute.ts (570 lines). snapshot-builder.ts is now 256 lines
containing only types, state prep, transcript fetching, and write logic.
